Added ignoreDiv option to skip headings within divs with a specific class.
This commit introduces a new option, ignoreDiv, to tocbot. When this option is set, any div with the specified class will be skipped, and its headings will not be added to the table of contents. This allows for more fine-grained control over which headings are included in the table of contents.

The following changes were made:

Added the ignoreDiv option to src/js/default-options.js.
Modified src/js/parse-content.js to check for the ignoreDiv class on parent divs of headings and skip them accordingly.